backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.47.2
    # via -r requirements.in
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
attrs==25.1.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.13.3
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
    #   selenium
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
distro==1.9.0
    # via anthropic
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==5.0.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via anthropic
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
jiter==0.8.2
    # via anthropic
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==24.2
    # via webdriver-manager
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.6
    # via anthropic
pydantic-core==2.27.2
    # via pydantic
pysocks==1.7.1
    # via urllib3
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   webdriver-manager
requests==2.32.3
    # via
    #   -r requirements.in
    #   webdriver-manager
selenium==4.29.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.6
    # via beautifulsoup4
trio==0.29.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   selenium
urllib3[socks]==2.3.0
    # via
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via -r requirements.in
websocket-client==1.8.0
    # via selenium
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wsproto==1.2.0
    # via trio-websocket
